Start working on shortcuts
